src/requirements.txt
einops==0.4.0
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.1
QtPy==2.0.1
six==1.16.0
torch==1.10.2
timm==0.5.4
torchaudio==0.10.2
torchvision==0.11.3
typing_extensions==4.1.1
vit-pytorch==0.26.7
tqdm
matplotlib
